Update ExcelDAO.java
Actualizado tipo de dato cédula en ExcelDAO

+++ src/java/com/example/dao/ExcelDAO.java
@@ -7,6 +7,6 @@ public interface ExcelDAO  extends AutoCloseable {
     public void create(Estudiante est) throws DAOException;//CREATE
     public Estudiante findByCedula(long cedula) throws DAOException;//READ
     public void update(Estudiante est) throws DAOException;//UPDATE
-    public void delete(int cedula) throws DAOException;//DELETE
+    public void delete(long cedula) throws DAOException;//DELETE
     public List<Estudiante> getAllEstudiantes() throws DAOException;
 }
